Move setupMounts and runSetupBuiltinVolumes to run_common.go
